pyproject.toml
[project]
name = "version-2-sequential-website-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.3.0",
    "litellm>=1.0.0",
    "duckduckgo-search>=6.0.0",
    "rich>=14.0.0",
]
